Remove StateTransitionOutput
