+++ full-node/sov-stf-runner/src/lib.rs
@@ -15,7 +15,7 @@ use std::path::Path;
 
 #[cfg(feature = "native")]
 use anyhow::Context;
-use borsh::{BorshDeserialize, BorshSerialize};
+
 #[cfg(feature = "native")]
 pub use config::RpcConfig;
 #[cfg(feature = "native")]
@@ -26,36 +26,15 @@ mod runner;
 pub use config::{from_toml_path, ProverServiceConfig, RollupConfig, RunnerConfig, StorageConfig};
 #[cfg(feature = "native")]
 pub use runner::*;
-use serde::de::DeserializeOwned;
-use serde::{Deserialize, Serialize};
-use sov_rollup_interface::da::DaSpec;
 
 /// Implements the `StateTransitionVerifier` type for checking the validity of a state transition
 pub mod verifier;
 
-#[derive(Serialize, BorshDeserialize, BorshSerialize, Deserialize)]
-// Prevent serde from generating spurious trait bounds. The correct serde bounds are already enforced by the
-// StateTransitionFunction, DA, and Zkvm traits.
-#[serde(bound = "StateRoot: Serialize + DeserializeOwned, Witness: Serialize + DeserializeOwned")]
-/// Data required to verify a state transition.
-pub struct StateTransitionData<StateRoot, Witness, Da: DaSpec> {
-    /// The state root before the state transition
-    pub pre_state_root: StateRoot,
-    /// The header of the da block that is being processed
-    pub da_block_header: Da::BlockHeader,
-    /// The proof of inclusion for all blobs
-    pub inclusion_proof: Da::InclusionMultiProof,
-    /// The proof that the provided set of blobs is complete
-    pub completeness_proof: Da::CompletenessProof,
-    /// The blobs that are being processed
-    pub blobs: Vec<<Da as DaSpec>::BlobTransaction>,
-    /// The witness for the state transition
-    pub state_transition_witness: Witness,
-}
-
 #[cfg(feature = "native")]
 /// Reads json file.
-pub fn read_json_file<T: DeserializeOwned, P: AsRef<Path>>(path: P) -> anyhow::Result<T> {
+pub fn read_json_file<T: serde::de::DeserializeOwned, P: AsRef<Path>>(
+    path: P,
+) -> anyhow::Result<T> {
     let path_str = path.as_ref().display();
 
     let data = std::fs::read_to_string(&path)
