+++ full-node/sov-stf-runner/src/prover_service/mod.rs
@@ -4,9 +4,8 @@ pub use parallel::ParallelProverService;
 use serde::Serialize;
 use sov_rollup_interface::da::DaSpec;
 use sov_rollup_interface::services::da::DaService;
+use sov_rollup_interface::zk::StateTransitionData;
 use thiserror::Error;
-
-use crate::StateTransitionData;
 
 /// The possible configurations of the prover.
 pub enum RollupProverConfig {
