+++ full-node/sov-stf-runner/src/verifier.rs
@@ -1,24 +1,9 @@
 use std::marker::PhantomData;
 
-use serde::{Deserialize, Serialize};
 use sov_rollup_interface::da::{BlockHeaderTrait, DaVerifier};
 use sov_rollup_interface::stf::StateTransitionFunction;
+use sov_rollup_interface::zk::StateTransitionData;
 use sov_rollup_interface::zk::{StateTransition, Zkvm, ZkvmGuest};
-
-use crate::StateTransitionData;
-
-#[derive(Serialize, Deserialize)]
-/// Output of the verifier.
-pub struct StateTransitionOutput<StateRoot, SlotHash> {
-    /// The state root before the state transition
-    pub pre_state_root: StateRoot,
-    /// The state root after the state transition
-    pub post_state_root: StateRoot,
-    /// Da block hash
-    pub da_block_hash: SlotHash,
-    /// The block height.
-    pub height: u64,
-}
 
 /// Verifies a state transition
 pub struct StateTransitionVerifier<ST, Da, Zk>
@@ -65,14 +50,6 @@ where
             &validity_condition,
             &mut data.blobs,
         );
-
-        /*
-        let out = StateTransitionOutput {
-            pre_state_root: data.pre_state_root,
-            post_state_root: result.state_root,
-            da_block_hash: data.da_block_header.hash(),
-            height: data.da_block_header.height(),
-        };*/
 
         let out: StateTransition<Da::Spec, Vec<u8>, _> = StateTransition {
             initial_state_root: data.pre_state_root,
