+++ rollup-interface/src/state_machine/zk/mod.rs
@@ -153,3 +153,23 @@ pub trait Matches<T> {
     /// Check if two items are a match
     fn matches(&self, other: &T) -> bool;
 }
+
+#[derive(Serialize, BorshDeserialize, BorshSerialize, Deserialize)]
+// Prevent serde from generating spurious trait bounds. The correct serde bounds are already enforced by the
+// StateTransitionFunction, DA, and Zkvm traits.
+#[serde(bound = "StateRoot: Serialize + DeserializeOwned, Witness: Serialize + DeserializeOwned")]
+/// Data required to verify a state transition.
+pub struct StateTransitionData<StateRoot, Witness, Da: DaSpec> {
+    /// The state root before the state transition
+    pub pre_state_root: StateRoot,
+    /// The header of the da block that is being processed
+    pub da_block_header: Da::BlockHeader,
+    /// The proof of inclusion for all blobs
+    pub inclusion_proof: Da::InclusionMultiProof,
+    /// The proof that the provided set of blobs is complete
+    pub completeness_proof: Da::CompletenessProof,
+    /// The blobs that are being processed
+    pub blobs: Vec<<Da as DaSpec>::BlobTransaction>,
+    /// The witness for the state transition
+    pub state_transition_witness: Witness,
+}
